backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.3.2
alembic==1.13.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2
pdfplumber==0.10.3
python-multipart==0.0.6

# Machine Learning & Stats
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Security & Auth
python-jose==3.3.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
PyJWT==2.11.0

# Scheduling
APScheduler==3.10.4

# Email & Notifications
python-dotenv==1.0.0
requests==2.31.0


# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.12.0
flake8==6.1.0
bandit==1.7.5
email-validator==2.1.0
argon2-cffi==23.1.0